+++ tanjun/clients.py
@@ -451,9 +451,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         "_hooks",
         "_interaction_not_found",
         "_slash_hooks",
-        "_is_alive",
         "_is_closing",
         "_listeners",
+        "_loop",
         "_message_hooks",
         "_metadata",
         "_modules",
@@ -506,9 +506,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         self._hooks: typing.Optional[tanjun_abc.AnyHooks] = hooks.AnyHooks().set_on_parser_error(on_parser_error)
         self._interaction_not_found: typing.Optional[str] = "Command not found"
         self._slash_hooks: typing.Optional[tanjun_abc.SlashHooks] = None
-        self._is_alive = False
         self._is_closing = False
         self._listeners: dict[type[hikari.Event], list[injecting.SelfInjectingCallback[None]]] = {}
+        self._loop: typing.Optional[asyncio.AbstractEventLoop] = None
         self._message_hooks: typing.Optional[tanjun_abc.MessageHooks] = None
         self._metadata: dict[typing.Any, typing.Any] = {}
         self._modules: dict[str, types.ModuleType] = {}
@@ -816,7 +816,12 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
     @property
     def is_alive(self) -> bool:
         # <<inherited docstring from tanjun.abc.Client>>.
-        return self._is_alive
+        return self._loop is not None
+
+    @property
+    def loop(self) -> typing.Optional[asyncio.AbstractEventLoop]:
+        # <<inherited docstring from tanjun.abc.Client>>.
+        return self._loop
 
     @property
     def message_hooks(self) -> typing.Optional[tanjun_abc.MessageHooks]:
@@ -1271,10 +1276,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         if add_injector:
             self.set_type_dependency(type(component), lambda: component)
 
-        if self._is_alive:
-            asyncio.get_running_loop().create_task(
-                self.dispatch_client_callback(ClientCallbackNames.COMPONENT_ADDED, component)
-            )
+        if self._loop:
+            self._loop.create_task(component.open())
+            self._loop.create_task(self.dispatch_client_callback(ClientCallbackNames.COMPONENT_ADDED, component))
 
         return self
 
@@ -1291,8 +1295,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         del self._components[component.name]
         stored_component.unbind_client(self)
 
-        if self._is_alive:
-            asyncio.get_running_loop().create_task(
+        if self._loop:
+            self._loop.create_task(component.close())
+            self._loop.create_task(
                 self.dispatch_client_callback(ClientCallbackNames.COMPONENT_REMOVED, stored_component)
             )
 
@@ -1375,7 +1380,7 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         self: _ClientT, event_type: type[hikari.Event], callback: tanjun_abc.ListenerCallbackSig, /
     ) -> _ClientT:
         # <<inherited docstring from tanjun.abc.Client>>.
-        injected = injecting.SelfInjectingCallback[None](self, callback)
+        injected: injecting.SelfInjectingCallback[None] = injecting.SelfInjectingCallback(self, callback)
         try:
             if callback in self._listeners[event_type]:
                 return self
@@ -1385,7 +1390,7 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         except KeyError:
             self._listeners[event_type] = [injected]
 
-        if self._is_alive and self._events:
+        if self._loop and self._events:
             self._events.subscribe(event_type, injected.__call__)
 
         return self
@@ -1400,7 +1405,7 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         if not self._listeners[event_type]:
             del self._listeners[event_type]
 
-        if self._is_alive and self._events:
+        if self._loop and self._events:
             self._events.unsubscribe(event_type, registered_callback.__call__)
 
         return self
@@ -1577,7 +1582,7 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         RuntimeError
             If the client isn't running.
         """
-        if not self._is_alive:
+        if not self._loop:
             raise RuntimeError("Client isn't active")
 
         if self._is_closing:
@@ -1604,7 +1609,9 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         if deregister_listeners and self._server:
             self._server.set_listener(hikari.CommandInteraction, None)
 
-        self._is_alive = False
+        await asyncio.gather(*(component.close() for component in self._components.copy().values()))
+
+        self._loop = None
         await self.dispatch_client_callback(ClientCallbackNames.CLOSED)
         self._is_closing = False
 
@@ -1620,10 +1627,10 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         RuntimeError
             If the client is already active.
         """
-        if self._is_alive:
+        if self._loop:
             raise RuntimeError("Client is already alive")
 
-        self._is_alive = True
+        self._loop = asyncio.get_running_loop()
         self._is_closing = False
         await self.dispatch_client_callback(ClientCallbackNames.STARTING)
 
@@ -1639,6 +1646,8 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
 
             self._grab_mention_prefix = False
 
+        await asyncio.gather(*(component.open() for component in self._components.copy().values()))
+
         if register_listeners and self._events:
             if event_type := self._accepts.get_event_type():
                 self._events.subscribe(event_type, self.on_message_create_event)
@@ -1652,7 +1661,7 @@ class Client(injecting.InjectorClient, tanjun_abc.Client):
         if register_listeners and self._server:
             self._server.set_listener(hikari.CommandInteraction, self.on_interaction_create_request)
 
-        asyncio.get_running_loop().create_task(self.dispatch_client_callback(ClientCallbackNames.STARTED))
+        self._loop.create_task(self.dispatch_client_callback(ClientCallbackNames.STARTED))
 
     async def fetch_rest_application_id(self) -> hikari.Snowflake:
         """Fetch the application ID of the application this client is linked to.